+++ src/main/java/calculator/calculator/Calculator.java
@@ -1,6 +1,5 @@
 package calculator.calculator;
 
-import calculator.delimiter.Delimiter;
 import calculator.delimiter.DelimiterFactory;
 import calculator.delimiter.DelimiterUtils;
 import java.util.List;
