+++ src/main/java/calculator/calculator/Delimiter.java
@@ -1,4 +1,4 @@
-package calculator.delimiter;
+package calculator.calculator;
 
 import java.util.List;
